float32 로 변경

+++ lib/dataset/deepfashion2.py
@@ -215,8 +215,8 @@ class DeepFashion2Dataset(JointsDataset):
             self.cls_stat[cls] += 1
             self.sample_list_of_cls.append(cls)
 
-            joints_3d = np.zeros((self.num_joints, 3), dtype=np.float)
-            joints_3d_vis = np.zeros((self.num_joints, 3), dtype=np.float)
+            joints_3d = np.zeros((self.num_joints, 3), dtype=np.float32)
+            joints_3d_vis = np.zeros((self.num_joints, 3), dtype=np.float32)
             for ipt in range(self.num_joints):
                 joints_3d[ipt, 0] = obj['keypoints'][ipt * 3 + 0]
                 joints_3d[ipt, 1] = obj['keypoints'][ipt * 3 + 1]
@@ -301,9 +301,9 @@ class DeepFashion2Dataset(JointsDataset):
                 num_boxes = num_boxes + 1
 
                 center, scale = self._box2cs(box)
-                joints_3d = np.zeros((self.num_joints, 3), dtype=np.float)
+                joints_3d = np.zeros((self.num_joints, 3), dtype=np.float32)
                 joints_3d_vis = np.ones(
-                    (self.num_joints, 3), dtype=np.float)
+                    (self.num_joints, 3), dtype=np.float32)
                 kpt_db.append({
                     'image': img_name,
                     'center': center,
@@ -342,9 +342,9 @@ class DeepFashion2Dataset(JointsDataset):
                 num_boxes = num_boxes + 1
 
                 center, scale = self._box2cs(box)
-                joints_3d = np.zeros((self.num_joints, 3), dtype=np.float)
+                joints_3d = np.zeros((self.num_joints, 3), dtype=np.float32)
                 joints_3d_vis = np.ones(
-                    (self.num_joints, 3), dtype=np.float)
+                    (self.num_joints, 3), dtype=np.float32)
                 kpt_db.append({
                     'image': img_name,
                     'center': center,
@@ -496,7 +496,7 @@ class DeepFashion2Dataset(JointsDataset):
                 _key_points = np.array([img_kpts[k]['keypoints']
                                         for k in range(len(img_kpts))])
                 key_points = np.zeros(
-                    (_key_points.shape[0], self.num_joints * 3), dtype=np.float
+                    (_key_points.shape[0], self.num_joints * 3), dtype=np.float32
                 )
 
                 for ipt in range(self.num_joints):
